little discrepancy tables to help reconciliation
To clarify: with a [address | supplement] table, where supplement is any BONK not from the 39 NFT txt files 500B rule ("target").

A "supplemental" address is either (A) in this table; (B) not in this table.

(A) If they are in the bonk_review_labeled.csv table; then they are most likely over-allocated (received more than 39 nft target); in which case the over-allocation is explainable as 'supplemental'.
     If they are in this table but NOT over-allocated: then the math is [owed = (target + supplement) - claimed] which requires the [address | supplement] table.
     If (target + supplement) - claimed  < 0,  then you have unexplained over-allocation. (They owe you a refund or you move the loss).
     If it is > 0, you owe them that amount.

(B) If they are not in this table; then w/e supplemental they are owed never got to them; and owed = supplement.